requirements.txt
ecs-logging==2.0.0
GPUtil==1.4.0
h2o==3.42.0.3
huggingface-hub==0.12.1
matplotlib==3.7.0
multidict==6.0.4
numpy==1.23.4
pandarallel==1.6.3
pandas==1.5.1
Pillow==9.4.0
plotly==5.11.0
psutil==5.9.4
py-singleton==1.0.0
pyarrow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
s3fs==2023.3.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.13.8
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.1
transformers==4.22.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.17.6
vaderSentiment==3.3.2
wordcloud==1.8.2.2
wordninja==2.0.0
wrapt==1.15.0
yarl==1.8.2
ipykernel